make conversion helper package specific
avoid accidentally providing floats where precision matters in the
future
