+++ core/assets/currencies_test.go
@@ -7,7 +7,6 @@ import (
 	"chainlink/core/assets"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestAssets_NewLinkAndString(t *testing.T) {
@@ -125,30 +124,4 @@ func TestAssets_Eth_UnmarshalJsonError(t *testing.T) {
 
 	err = json.Unmarshal([]byte(`1`), &eth)
 	assert.Equal(t, assets.ErrNoQuotesForCurrency, err)
-}
-
-func TestAssets_Eth_Float64_Success(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		name      string
-		input     string
-		want      float64
-		wantError bool
-	}{
-		{"zero", "0", 0, false},
-		{"small", "1", 0.000000000000000001, false},
-		{"rounding", "12345678901234567890", 12.345678901234567, false},
-		{"large", "123456789012345678901234567890", 123456789012.34567, false},
-		{"extreme", "1234567890123456789012345678901234567890123456789012345678901234567890", 1.2345678901234568e+51, false},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			eth := assets.NewEth(0)
-			eth.SetString(test.input, 10)
-			float, err := eth.Float64()
-			require.NoError(t, err)
-			require.Equal(t, test.want, float)
-		})
-	}
 }
